+++ coinapi/src/main/java/pl/cyfronet/coin/api/beans/AtomicServiceRequest.java
@@ -34,8 +34,8 @@ public class AtomicServiceRequest {
 	private Boolean published;
 	
 	private Float cpu;
-	private Long memory;
-	private Long disk;
+	private Integer memory;
+	private Integer disk;
 	
 	public String getName() {
 		return name;
@@ -111,19 +111,19 @@ public class AtomicServiceRequest {
 		this.cpu = cpu;
 	}
 
-	public Long getMemory() {
+	public Integer getMemory() {
 		return memory;
 	}
 
-	public void setMemory(Long memory) {
+	public void setMemory(Integer memory) {
 		this.memory = memory;
 	}
 
-	public Long getDisk() {
+	public Integer getDisk() {
 		return disk;
 	}
 
-	public void setDisk(Long disk) {
+	public void setDisk(Integer disk) {
 		this.disk = disk;
 	}
