+++ coined/src/main/java/pl/cyfronet/coin/impl/air/client/AppliancePreferences.java
@@ -2,8 +2,8 @@ package pl.cyfronet.coin.impl.air.client;
 
 public class AppliancePreferences {
 	private Float cpu;
-	private Long memory;
-	private Long disk;
+	private Integer memory;
+	private Integer disk;
 
 	public Float getCpu() {
 		return cpu;
@@ -13,19 +13,19 @@ public class AppliancePreferences {
 		this.cpu = cpu;
 	}
 
-	public Long getMemory() {
+	public Integer getMemory() {
 		return memory;
 	}
 
-	public void setMemory(Long memory) {
+	public void setMemory(Integer memory) {
 		this.memory = memory;
 	}
 
-	public Long getDisk() {
+	public Integer getDisk() {
 		return disk;
 	}
 
-	public void setDisk(Long disk) {
+	public void setDisk(Integer disk) {
 		this.disk = disk;
 	}
